HomePage

+++ Radiostanciya/Controllers/GenresController.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Radiostanciya.Models;
+using Microsoft.Extensions.Caching.Memory;
 using Radiostanciya.ViewModels;
 using Radiostanciya.ViewModels.GenreViewModels;
 
@@ -15,16 +16,32 @@ namespace Radiostanciya.Controllers
     [Authorize]
     public class GenresController : Controller
     {
-        ApplContext db;
-        public GenresController(ApplContext db)
+        private readonly ApplContext db;
+        private readonly IMemoryCache memoryCache;
+        public GenresController(ApplContext db, IMemoryCache memoryCache)
         {
             this.db = db;
+            this.memoryCache = memoryCache;
         }
 
-        [ResponseCache(Duration = 2 * 15 + 240, Location = ResponseCacheLocation.Any)]
+        [ResponseCache(CacheProfileName = "Caching")]
         public async Task<IActionResult> Index(int? id, string name, int page = 0,
             SortState sortOrder = SortState.IdAsc)
         {
+            // Проверяем, есть ли данные в кэше
+            if (!memoryCache.TryGetValue("GenresList", out List<Genre> genres))
+            {
+                // Данные отсутствуют в кэше, извлекаем из базы данных
+                genres = await db.Genres.ToListAsync();
+
+                // Кэшируем данные
+                memoryCache.Set("GenresList", genres, new MemoryCacheEntryOptions
+                {
+                    AbsoluteExpirationRelativeToNow = TimeSpan.FromMinutes(10), // Кэшировать на 10 минут
+                    SlidingExpiration = TimeSpan.FromMinutes(2) // Обновлять, если не использовалось более 2 минут
+                });
+            }
+
             ViewData["IsAdmin"] = User.IsInRole("Admin");
 
             int pageSize = 10;  // количество элементов на странице
@@ -77,12 +94,23 @@ namespace Radiostanciya.Controllers
                 HttpContext.Response.Cookies.Append("selectedId", id.Value.ToString());
             }
 
-            return View(viewModel);
+            /*// Вывести данные из кэша в лог
+            if (memoryCache.TryGetValue("GenresList", out List<Genre> cachedGenres))
+            {
+                foreach (var genre in cachedGenres)
+                {
+                    Console.WriteLine($"Cached Genre: {genre.Id} - {genre.Name}");
+                }
+            }*/
+            return await Task.FromResult(View(viewModel));
         }
 
         [HttpGet]
         public ActionResult Insert(string name, string description)
         {
+            // Обновляем кэш при вставке новых данных
+            memoryCache.Remove("GenresList");
+
             Genre genre = new Genre
             {
                 Name = name,
@@ -97,14 +125,21 @@ namespace Radiostanciya.Controllers
         [HttpGet]
         public ActionResult Delete(int id)
         {
+            // Обновляем кэш при удалении данных
+            memoryCache.Remove("GenresList");
+
             Genre genre = null;
             try
             {
-                genre = db.Genres.Where(c => c.Id == id).First();
-                db.Genres.Remove(genre);
-                db.SaveChanges();
+                genre = db.Genres.Where(c => c.Id == id).FirstOrDefault();
+                if (genre != null)
+                {
+                    db.Genres.Remove(genre);
+                    db.SaveChanges();
+                }
             }
             catch { }
+
             JsonResult data = Json(genre);
             return data;
         }
@@ -112,15 +147,22 @@ namespace Radiostanciya.Controllers
         [HttpGet]
         public ActionResult Update(int id, string name, string desc)
         {
+            // Обновляем кэш при обновлении данных
+            memoryCache.Remove("GenresList");
+
             Genre genre = null;
             try
             {
-                genre = db.Genres.Where(c => c.Id == id).First();
-                genre.Name = name;
-                genre.Description = desc;
-                db.SaveChanges();
+                genre = db.Genres.Where(c => c.Id == id).FirstOrDefault();
+                if (genre != null)
+                {
+                    genre.Name = name;
+                    genre.Description = desc;
+                    db.SaveChanges();
+                }
             }
             catch { }
+
             JsonResult data = Json(desc);
             return data;
         }
